requirements.txt
# Web framework and ASGI server
fastapi[standard]
starlette
uvicorn

# Data and Database
pydantic
beanie
motor
pymongo

# AI and Language Models
openai
langgraph
langchain
langchain-openai
langchain-community
langchain-core

# Fetcher Sources
arxiv
wikipedia
duckduckgo-search
ddgs

# Unit Tests
httpx
pytest
pytest-asyncio

# Utilities
xmltodict
pymupdf
python-dotenv